+++ react-components/src/components/Dashboard/index.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/prefer-default-export */
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 
@@ -11,39 +12,65 @@ export const Dashboard = ({ sections }) => {
     setModal(true)
     // record click on disable section
     analytics({
-      'event':'ctaFeature',
-      'featureTitle':e.target.dataset.sectiontitle
+      event: 'ctaFeature',
+      featureTitle: e.target.dataset.sectiontitle,
     })
   }
 
   return (
     <>
-      <ComingSoon
-        onClick={() => setModal(false)}
-        isOpen={modal}
-      />
+      <ComingSoon onClick={() => setModal(false)} isOpen={modal} />
       {sections.map(({ title, url, disabled }) => (
-        <div className='c-1-3' key={url}>
-          <div className='bg-white m-b-s section-list__item'>
-            {disabled ?
-              <div className='w-full link section-list__disabled' role='button' onClick={openComingSoonModal} >
-                <div className='bg-blue-deep-20'>
-                  <span className='section-list__coming bg-blue-deep-80 text-white body-m p-xxs' data-sectiontitle={title}>Coming soon</span>
-                  <img className='w-full p-h-s p-t-m p-b-s' src='/static/images/coming-soon.svg' alt={title}  data-sectiontitle={title}/>
+        <div className="c-1-3-xl c-1-2-m" key={url}>
+          <div className="bg-white m-b-s section-list__item">
+            {disabled ? (
+              <div
+                className="w-full link section-list__disabled section-list__link"
+                onClick={openComingSoonModal}
+                aria-hidden="true"
+                role="button"
+              >
+                <div className="bg-blue-deep-20 section-list__image-container">
+                  <span
+                    className="section-list__coming bg-blue-deep-80 text-white body-m p-xxs"
+                    data-sectiontitle={title}
+                  >
+                    Coming soon
+                  </span>
+                  <img
+                    className="w-full p-h-s p-t-m p-b-s"
+                    src="/static/images/coming-soon.svg"
+                    alt={title}
+                    data-sectiontitle={title}
+                  />
                 </div>
-                <div className='p-v-s p-h-xs'>
-                  <h3 className='h-xs text-blue-deep-80 p-0' data-sectiontitle={title}>{title}</h3>
+                <div className="p-v-s p-h-xs">
+                  <h3
+                    className="h-xs text-blue-deep-80 p-0"
+                    data-sectiontitle={title}
+                  >
+                    {title}
+                  </h3>
                 </div>
-              </div> :
-              <a className='w-full link' href={url} title={title}>
-                <div className='bg-blue-deep-20'>
-                  <img className='w-full p-h-s p-t-m p-b-s' src='/static/images/ep-placeholder.svg' alt={title} />
+              </div>
+            ) : (
+              <a
+                className="w-full link section-list__link"
+                href={url}
+                title={title}
+              >
+                <div className="bg-blue-deep-20 section-list__image-container">
+                  <img
+                    className="w-full p-h-s p-t-m p-b-s"
+                    src="/static/images/ep-placeholder.svg"
+                    alt={title}
+                  />
                 </div>
-                <div className='p-v-s p-h-xs'>
-                  <h3 className='h-xs text-blue-deep-80 p-0'>{title}</h3>
+                <div className="p-v-s p-h-xs">
+                  <h3 className="h-xs text-blue-deep-80 p-0">{title}</h3>
                 </div>
               </a>
-            }
+            )}
           </div>
         </div>
       ))}
@@ -56,8 +83,7 @@ Dashboard.propTypes = {
     PropTypes.shape({
       title: PropTypes.string,
       url: PropTypes.string,
-      disabled: PropTypes.bool
+      disabled: PropTypes.bool,
     })
   ).isRequired,
 }
-