+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 
 ### Fixed bugs
 
+
+## [1.0.1](https://github.com/uktrade/great-cms/releases/tag/1.0.1)
+[Full Changelog](https://github.com/uktrade/great-cms/compare/1.0.0...1.0.1)
+
+### Implemented enhancements
+
+### Fixed bugs
+
+- GP2-1158 - Verification Code modal
+
 ## [1.0.0](https://github.com/uktrade/great-cms/releases/tag/1.0.0)
 [Full Changelog](https://github.com/uktrade/great-cms/releases/tag/1.0.0)